refactor: Rename files for improved readability and structure
- Renamed `note_type_model.dart` to `notetype_model.dart` for consistency.
- Renamed `add_note_type_screen.dart` to `notetype_form_screen.dart` to reflect shared resource functionality.
- Renamed `add_note_screen.dart` to `note_form_screen.dart` for clarity in form usage.
- Updated import statements across the codebase to match the new file names.
