+++ lib/notetype_model.dart
